move debugging lines to a separate class
