+++ src/dfastbe/bank_erosion/bank_erosion.py
@@ -40,6 +40,7 @@ from dfastbe.kernel import get_km_bins, moving_avg, comp_erosion_eq, comp_erosio
                             get_zoom_extends
 from dfastbe.support import move_line
 from dfastbe import plotting as df_plt
+from dfastbe.bank_erosion.debugger import Debugger
 from dfastbe.io import (
     LineGeometry, ConfigFile,
     log_text,
@@ -79,6 +80,7 @@ class Erosion:
         self.simulation_data = self.river_data.simulation_data()
         self.sim_files, self.p_discharge = self.river_data.get_erosion_sim_data(self.river_data.num_discharge_levels)
         self.bl_processor = BankLinesProcessor(self.river_data)
+        self.debugger = Debugger(config_file, self.river_data)
 
     @property
     def config_file(self) -> ConfigFile:
@@ -313,81 +315,6 @@ class Erosion:
             zfw_ini.append(simulation_data.water_level_face[ii])
         fairway_data.fairway_initial_water_levels = zfw_ini
 
-    def debug_process_discharge_levels_1(
-        self, ib, bank_data, fairway_data, erosion_inputs, pars, hfw, dn_eq1, dv_eq1, bcrds, bank_height,
-        line_size
-    ):
-        bcrds_mid = (bcrds[:-1] + bcrds[1:]) / 2
-        bank_coords_points = [Point(xy) for xy in bcrds_mid]
-        bank_coords_geo = GeoSeries(
-            bank_coords_points, crs=self.config_file.crs
-        )
-        params = {
-            "chainage": bank_data.bank_chainage_midpoints[ib],
-            "x": bcrds_mid[:, 0],
-            "y": bcrds_mid[:, 1],
-            "iface_fw": bank_data.fairway_face_indices[ib],
-            "iface_bank": bank_data.bank_face_indices[ib],  # bank_index
-            "zb": bank_height[ib],
-            "len": line_size[ib],
-            "zw0": fairway_data.fairway_initial_water_levels[ib],
-            "vship": pars["v_ship"][ib],
-            "shiptype": pars["ship_type"][ib],
-            "draught": pars["t_ship"][ib],
-            "mu_slp": pars["mu_slope"][ib],
-            "dist_fw": bank_data.fairway_distances[ib],
-            "dfw0": erosion_inputs.wave_fairway_distance_0[ib],
-            "dfw1": erosion_inputs.wave_fairway_distance_1[ib],
-            "hfw": hfw,
-            "zss": erosion_inputs.bank_protection_level[ib],
-            "dn": dn_eq1,
-            "dv": dv_eq1,
-        }
-
-        write_shp(bank_coords_geo, params, f"{str(self.river_data.output_dir)}{os.sep}debug.EQ.B{ib + 1}.shp")
-        write_csv(params, f"{str(self.river_data.output_dir)}{os.sep}debug.EQ.B{ib + 1}.csv")
-
-    def debug_process_discharge_levels_2(
-        self, ib, iq, bank_data, fairway_data, erosion_inputs, pars, hfw, bcrds, velocity, bank_height, line_size,
-        water_level, chezy, dniqib, dviqib, dn_ship, dn_flow
-    ):
-        bcrds_mid = (bcrds[:-1] + bcrds[1:]) / 2
-
-        bank_coords_points = [Point(xy1) for xy1 in bcrds_mid]
-        bank_coords_geo = GeoSeries(bank_coords_points, crs=self.config_file.crs)
-        params = {
-            "chainage": bank_data.bank_chainage_midpoints[ib],
-            "x": bcrds_mid[:, 0],
-            "y": bcrds_mid[:, 1],
-            "iface_fw": bank_data.fairway_face_indices[ib],
-            "iface_bank": bank_data.bank_face_indices[ib],  # bank_index
-            "u": velocity[iq][ib],
-            "zb": bank_height[ib],
-            "len": line_size[ib],
-            "zw": water_level[iq][ib],
-            "zw0": fairway_data.fairway_initial_water_levels[ib],
-            "tauc": erosion_inputs.tauc[ib],
-            "nship": pars["n_ship"][ib],
-            "vship": pars["v_ship"][ib],
-            "nwave": pars["n_wave"][ib],
-            "shiptype": pars["ship_type"][ib],
-            "draught": pars["t_ship"][ib],
-            "mu_slp": pars["mu_slope"][ib],
-            "mu_reed": pars["mu_reed"][ib],
-            "dist_fw": bank_data.fairway_distances[ib],
-            "dfw0": erosion_inputs.wave_fairway_distance_0[ib],
-            "dfw1": erosion_inputs.wave_fairway_distance_1[ib],
-            "hfw": hfw,
-            "chez": chezy[iq][ib],
-            "zss": erosion_inputs.bank_protection_level[ib],
-            "dn": dniqib,
-            "dv": dviqib,
-            "dnship": dn_ship,
-            "dnflow": dn_flow,
-        }
-        write_shp(bank_coords_geo, params, f"{str(self.river_data.output_dir)}{os.sep}debug.Q{iq + 1}.B{ib + 1}.shp")
-        write_csv(params, f"{str(self.river_data.output_dir)}{os.sep}debug.Q{iq + 1}.B{ib + 1}.csv")
-
     def _prepare_initial_conditions(
         self, config_file: ConfigFile, num_stations_per_bank: List[int], fairway_data: FairwayData
     ) -> ErosionInputs:
@@ -499,7 +426,6 @@ class Erosion:
                 "discharge_header",
                 data={"i": iq + 1, "p": self.p_discharge[iq], "t": self.p_discharge[iq] * self.river_data.erosion_time,},
             )
-
             iq_str = f"{iq + 1}"
 
             log_text("read_q_params", indent="  ")
@@ -604,8 +530,8 @@ class Erosion:
                     eq_eroded_vol.append(dv_eq1)
 
                     if self.river_data.debug:
-                        self.debug_process_discharge_levels_1(
-                            ib, bank_data, fairway_data, erosion_inputs, pars, hfw, dn_eq1, dv_eq1, bcrds,
+                        self.debugger.debug_process_discharge_levels_1(
+                            ind, bank_data, fairway_data, erosion_inputs, pars, hfw, dn_eq1, dv_eq1, coords_i,
                             bank_height, line_size
                         )
 
@@ -637,9 +563,9 @@ class Erosion:
                 ship_wave_min[iq].append(ship_wave_min_ib)
 
                 if self.river_data.debug:
-                    self.debug_process_discharge_levels_2(
-                        ib, iq, bank_data, fairway_data, erosion_inputs, pars, hfw, bcrds, velocity, bank_height, line_size,
-                        water_level, chezy, dniqib, dviqib, dn_ship, dn_flow
+                    self.debugger.debug_process_discharge_levels_2(
+                        ind, iq, bank_data, fairway_data, erosion_inputs, pars, hfw, coords_i, velocity_iq, bank_height, line_size,
+                        water_level_iq, chezy_iq, dn_tot, dv_tot, dn_ship, dn_flow
                     )
 
                 # shift bank lines
